+++ src/main/java/org/folio/edge/resolver/Constants.java
@@ -4,6 +4,9 @@ public final class Constants {
   public static final String PATH_PARAM_USER_ID = "userId";
   public static final String PATH_PARAM_LOAN_ID = "loanId";
   public static final String PATH_PARAM_REQUEST_ID = "requestId";
+  public static final String PATH_PARAM_INSTANCE_ID = "instanceId";
+  public static final String PATH_PARAM_HOLDING_ID = "holdingId";
+  public static final String PATH_PARAM_ITEM_ID = "itemId";
 
   private Constants() {
     throw new AssertionError("Cannot construct");