+++ NEWS.md
@@ -1,4 +1,5 @@
 ## 1.0.0 Unreleased
+ * [EDGRESOLV-4](https://issues.folio.org/browse/EDGRESOLV-4): Added instance, holding and item UUID resolution APIs
  * [EDGRESOLV-3](https://issues.folio.org/browse/EDGRESOLV-3): Added loan and request UUID resolution APIs
  * [EDGRESOLV-2](https://issues.folio.org/browse/EDGRESOLV-2): Added user UUID resolution API
  * [EDGRESOLV-1](https://issues.folio.org/browse/EDGRESOLV-1): Initial commit of boiler plate